World Manager updates

+++ MCServerlauncher/MCServerlauncher/Diags/WorldMod.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.IO;
 using System.IO.Compression;
+using System.Diagnostics;
 
 namespace MCServerLauncher
 {
@@ -76,29 +77,40 @@ namespace MCServerLauncher
         private void WorldMod_DragDrop(object sender, DragEventArgs e)
         {
             string[] fileList = (string[])e.Data.GetData(DataFormats.FileDrop, false);
-            ZipArchive za = ZipFile.Open(fileList[0], ZipArchiveMode.Read);
-            var tempp = Path.GetTempPath() + "MCServerLauncher";
-            MessageBox.Show(tempp);
-            if(!Directory.Exists(tempp))
-                Directory.CreateDirectory(tempp);
 
-            za.ExtractToDirectory(tempp);
-            za.Dispose();
-
-            string[] dirs = Directory.GetDirectories(tempp);
-
-            foreach(string dir in dirs)
+            if (fileList[0].Contains(".zip"))
             {
-                if(File.Exists(dir + "/level.dat"))
+                ZipArchive za = ZipFile.Open(fileList[0], ZipArchiveMode.Read);
+                var tempp = Path.GetTempPath() + "MCServerLauncher";
+                MessageBox.Show(tempp);
+                if (!Directory.Exists(tempp))
+                    Directory.CreateDirectory(tempp);
+
+                za.ExtractToDirectory(tempp);
+                za.Dispose();
+
+                string[] dirs = Directory.GetDirectories(tempp);
+
+                foreach (string dir in dirs)
                 {
-                    var ds = dir.Split('\\');
-                    Directory.Move(dir, svdir + "/" + ds[ds.Length - 1]);
-                   
-                    //svp.ServerVars["level-name"].value = ds[ds.Length - 1];
-                    //lblCurrent.Text = "Current World: " + ds[ds.Length - 1];
-                    //File.WriteAllText(svdir + "/server.properties", svp.ToString());
-                    RefreshWorlds();
+                    if (File.Exists(dir + "/level.dat"))
+                    {
+                        var ds = dir.Split('\\');
+                        Directory.Move(dir, svdir + "/" + ds[ds.Length - 1]);
+
+                        //svp.ServerVars["level-name"].value = ds[ds.Length - 1];
+                        //lblCurrent.Text = "Current World: " + ds[ds.Length - 1];
+                        //File.WriteAllText(svdir + "/server.properties", svp.ToString());
+                        RefreshWorlds();
+                    }
                 }
+
+            } else if(Directory.Exists(fileList[0]))
+                if (File.Exists(fileList[0] + "/level.dat"))
+            {
+                var ds = fileList[0].Split('\\');
+                Directory.Move(fileList[0], svdir + "/" + ds[ds.Length - 1]);
+                    Refresh();
             }
         }
 
@@ -112,7 +124,7 @@ namespace MCServerLauncher
 
         private void LstWorlds_DragDrop(object sender, DragEventArgs e)
         {
-
+            WorldMod_DragDrop(sender, e);
         }
 
 
@@ -122,9 +134,73 @@ namespace MCServerLauncher
         }
         private void BtnSet_Click(object sender, EventArgs e)
         {
+            if (lstWorlds.SelectedIndex == -1)
+            {
+                MessageBox.Show("Please select a world first!");
+                return;
+            }
+
             SetWorldName(lstWorlds.Items[lstWorlds.SelectedIndex].ToString());
             svp.ServerVars["level-name"].value = lstWorlds.Items[lstWorlds.SelectedIndex].ToString();
             File.WriteAllText(svdir + "/server.properties", svp.ToString());
+        }
+
+        private void BtnExit_Click(object sender, EventArgs e)
+        {
+            this.Close();
+        }
+
+        private void LstWorlds_DragEnter(object sender, DragEventArgs e)
+        {
+            if (e.Data.GetDataPresent(DataFormats.FileDrop))
+                e.Effect = DragDropEffects.Copy;
+            else
+                e.Effect = DragDropEffects.None;
+        }
+
+
+        void THANOS_SNAP(string dir)
+        {
+
+            foreach (string file in Directory.GetFiles(dir))
+                File.Delete(file);
+            foreach (string sdir in Directory.GetDirectories(dir))
+            {
+                THANOS_SNAP(sdir);
+                Directory.Delete(sdir);
+            }
+
+        }
+        private void BtnDelete_Click(object sender, EventArgs e)
+        {
+            
+            if(lstWorlds.SelectedIndex == -1)
+            {
+                MessageBox.Show("Please select a world first!");
+                return;
+            }
+            DialogResult dg = MessageBox.Show("Are you sure you want to delete \"" + lstWorlds.Items[lstWorlds.SelectedIndex].ToString() + "\"?\nThis cannot be undone", "Confirm.", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+            if (dg == DialogResult.Yes)
+            {
+                THANOS_SNAP(svdir + "/" + lstWorlds.Items[lstWorlds.SelectedIndex].ToString());
+                Directory.Delete(svdir + "/" + lstWorlds.Items[lstWorlds.SelectedIndex].ToString());
+                RefreshWorlds();
+            }
+            else
+                return;
+        }
+
+        private void BtnOpen_Click(object sender, EventArgs e)
+        {
+            if (lstWorlds.SelectedIndex == -1)
+            {
+                MessageBox.Show("Please select a world first!");
+                return;
+            }
+            var dir = "\"" + Directory.GetCurrentDirectory() + "\\" + svdir + "\\" + lstWorlds.Items[lstWorlds.SelectedIndex].ToString() + "\"";
+            ProcessStartInfo a = new ProcessStartInfo("explorer", dir);
+            System.Diagnostics.Process.Start(a);
         }
     }
 }
